save state #6

+++ patch/TAKReversed/src/GlobalPointers/GlobalPointers.h
@@ -3,9 +3,9 @@
 
 namespace GlobalPointers
 {
-	const DWORD ptr_22D55C = 0x22D55C;
+	const uintptr_t ptr_22D55C = 0x22D55C;
 
 	// 65DDB4
-	const DWORD GameInterfaceHandler = 0x25DDB4;
-	const DWORD ptr_241A48 = 0x241A48;
+	const uintptr_t GameInterfaceHandler = 0x25DDB4;
+	const uintptr_t ptr_241A48 = 0x241A48;
 }
